Fixing posts errors
