+++ emendation_box/urls.py
@@ -1,6 +1,7 @@
 from .views import EmendationBoxListViewSet
 from .views import EmendationBoxTypeListViewSet
 from .views import EmendationBoxStructureListViewSet
+from .views import PostListViewSet
 
 
 # from rest_framework import renderers
@@ -24,6 +25,18 @@ emendation_box_type = EmendationBoxTypeListViewSet.as_view({
 })
 
 emendation_box_structure = EmendationBoxStructureListViewSet.as_view({
+    'get': 'list',
+    'patch': 'partial_update',
+    'put': 'update',
+    'delete': 'destroy',
+})
+
+post = PostListViewSet.as_view({
+    'get': 'list',
+    'post': 'create',
+})
+
+post_detail = PostListViewSet.as_view({
     'get': 'list',
     'patch': 'partial_update',
     'put': 'update',
