+++ emendation_box/views.py
@@ -3,6 +3,8 @@ from .serializers import EmendationBoxSerializer
 from .serializers import EmendationBoxTypeSerializer
 from .serializers import EmendationBoxStructureSerializer
 from .models import EmendationBoxStructure, EmendationBoxType, EmendationBox
+from .serializers import PostSerializer
+from .models import Post
 
 
 # Create your views here.
@@ -19,3 +21,8 @@ class EmendationBoxTypeListViewSet(viewsets.ModelViewSet):
 class EmendationBoxStructureListViewSet(viewsets.ModelViewSet):
     queryset = EmendationBoxStructure.objects.all()
     serializer_class = EmendationBoxStructureSerializer
+
+
+class PostListViewSet(viewsets.ModelViewSet):
+    queryset = Post.objects.all()
+    serializer_class = PostSerializer
