+++ sigi_op/settings.py
@@ -47,7 +47,6 @@ INSTALLED_APPS = [
     'emendation_box',
     'uplink',
     'technical_reserve',
-    'post',
 ]
 
 TEST_RUNNER = 'django_nose.NoseTestSuiteRunner'
